Move scope functions to a separate file
