+++ src/index.js
@@ -2,6 +2,7 @@ const jaystring = require('./jaystring')
 const genfun = require('./generate-function')
 const { toPointer, resolveReference, joinPath } = require('./pointer')
 const formats = require('./formats')
+const functions = require('./scope-functions')
 const KNOWN_KEYWORDS = require('./known-keywords')
 
 // name is assumed to be already processed and can contain complex paths
@@ -19,70 +20,6 @@ types.object = (name) => `typeof ${name} === "object" && ${name} && !Array.isArr
 types.number = (name) => `typeof ${name} === "number"`
 types.integer = (name) => `Number.isInteger(${name})`
 types.string = (name) => `typeof ${name} === "string"`
-
-const unique = (array) => {
-  const objects = []
-  const primitives = new Set()
-  let primitivesCount = 0
-  for (const item of array) {
-    if (typeof item === 'object') {
-      objects.push(item)
-    } else {
-      primitives.add(item)
-      if (primitives.size !== ++primitivesCount) return false
-    }
-  }
-  for (let i = 1; i < objects.length; i++)
-    for (let j = 0; j < i; j++) if (deepEqual(objects[i], objects[j])) return false
-  return true
-}
-
-const isMultipleOf = (value, multipleOf) => {
-  if (typeof multipleOf !== 'number' || !Number.isFinite(value))
-    throw new Error('multipleOf is not a number')
-  if (typeof value !== 'number' || !Number.isFinite(value)) return false
-  if (value === 0) return true
-  if (multipleOf === 0) return false
-  const digitsAfterDot = (number) => {
-    if ((number | 0) === number) return 0
-    return String(number)
-      .split('.')
-      .pop().length
-  }
-  const digits = digitsAfterDot(multipleOf)
-  if (digits === 0) return value % multipleOf === 0
-  const valueDigits = digitsAfterDot(value)
-  if (valueDigits > digits) return false
-  const factor = Math.pow(10, digits)
-  return Math.round(factor * value) % Math.round(factor * multipleOf) === 0
-}
-
-// supports only JSON-stringifyable objects, defaults to false for unsupported
-// also uses ===, not Object.is, i.e. 0 === -0, NaN !== NaN
-// symbols and non-enumerable properties are ignored!
-const deepEqual = (obj, obj2) => {
-  if (obj === obj2) return true
-  if (!obj || !obj2 || typeof obj !== typeof obj2) return false
-
-  const proto = Object.getPrototypeOf(obj)
-  if (proto !== Object.getPrototypeOf(obj2)) return false
-
-  if (proto === Array.prototype) {
-    if (!Array.isArray(obj) || !Array.isArray(obj2)) return false
-    if (obj.length !== obj2.length) return false
-    return obj.every((x, i) => deepEqual(x, obj2[i]))
-  } else if (proto === Object.prototype) {
-    const [keys, keys2] = [Object.keys(obj), Object.keys(obj2)]
-    if (keys.length !== keys2.length) return false
-    const keyset2 = new Set(keys2)
-    return keys.every((key) => keyset2.has(key) && deepEqual(obj[key], obj2[key]))
-  }
-  return false
-}
-
-// for correct Unicode code points processing
-// https://mathiasbynens.be/notes/javascript-unicode#accounting-for-astral-symbols
-const stringLength = (string) => [...string].length
 
 const scopeSyms = Symbol('syms')
 const scopeRefCache = Symbol('refcache')
@@ -353,7 +290,7 @@ const compile = (schema, root, opts, scope, basePathRoot) => {
 
     const makeCompare = (name, complex) => {
       if (complex) {
-        scope.deepEqual = deepEqual
+        scope.deepEqual = functions.deepEqual
         return (e) => `deepEqual(${name}, ${JSON.stringify(e)})`
       }
       return (e) => `(${name} === ${JSON.stringify(e)})`
@@ -397,7 +334,7 @@ const compile = (schema, root, opts, scope, basePathRoot) => {
       validateTypeApplicable('number', 'integer')
       if (type !== 'number' && type !== 'integer') fun.write('if (%s) {', types.number(name))
 
-      scope.isMultipleOf = isMultipleOf
+      scope.isMultipleOf = functions.isMultipleOf
       fun.write('if (!isMultipleOf(%s, %d)) {', name, node[multipleOf])
 
       error('has a remainder')
@@ -414,7 +351,7 @@ const compile = (schema, root, opts, scope, basePathRoot) => {
       validateTypeApplicable('string')
       if (type !== 'string') fun.write('if (%s) {', types.string(name))
 
-      scope.stringLength = stringLength
+      scope.stringLength = functions.stringLength
       fun.write('if (stringLength(%s) > %d) {', name, node.maxLength)
       error('has longer length than allowed')
       fun.write('}')
@@ -428,7 +365,7 @@ const compile = (schema, root, opts, scope, basePathRoot) => {
       validateTypeApplicable('string')
       if (type !== 'string') fun.write('if (%s) {', types.string(name))
 
-      scope.stringLength = stringLength
+      scope.stringLength = functions.stringLength
       fun.write('if (stringLength(%s) < %d) {', name, node.minLength)
       error('has less length than allowed')
       fun.write('}')
@@ -598,8 +535,8 @@ const compile = (schema, root, opts, scope, basePathRoot) => {
     if (node.uniqueItems === true) {
       validateTypeApplicable('array')
       if (type !== 'array') fun.write('if (%s) {', types.array(name))
-      scope.unique = unique
-      scope.deepEqual = deepEqual
+      scope.unique = functions.unique
+      scope.deepEqual = functions.deepEqual
       fun.write('if (!(unique(%s))) {', name)
       error('must be unique')
       fun.write('}')
